Master restoration (#4)
* change repo config to github repo

* import visualizer models with git lfs

* Docker startup script updates
- Updated auv-docker.py for NVIDIA Jetson container
- Docker image startup for x64 vs. aarch64 based machines

* Small auv-docker.py fix for proper clize parsing

* Aaron's webserver updates from 2022-2024

* Infrastructure Updates from 2022-2024
- Massive Dockerfile overhaul for NVIDIA Drivers
- Moved unused systems to deprecated folder
- configure.py retiring deprecated systems
- Additional cleanup

* Add .pt to git lfs

* Missions from 2022-2024

* Vision modules 2022-2024

* Serial Updates 2022-2024
- Infrastructure tweaking in build files
- BMB poda and podb initial creation

* Configuration updates 2022-2024

* Install directory updates 2022-2024
- ZED integration into Dockerfile
- The great CUDA struggle in 2023

* Control updates 2022-2024
- Actuator helm
- Tweaks to thruster helm

* Hydrocode Updates 2022-2024

* Sensor updates 2022-2024
- Jeffrey's 2023 Kalman rewrite (is it being used?)
- Changing numpy to cupy

* LED Updates 2022-2024

* Trogdor Updates 2022-2024 (updated to Polaris)

* Miscellaneous updates

* Visualizer overhaul 2022-2024
- Just a lot of changes to the visualizer, currently has the 2023 simulator

* Self-test update 2024

* removed remote webserver

* updated github url

* changed mounting local .ssh keys (sus)

* Final ssh mount change

+++ mission/runner.py
@@ -89,7 +89,8 @@ def run(mission : Coroutine[Any, Any, None], name : str):
         sys.exit(1)
 
     def release_lock():
-        os.rmdir(lock_dir)
+        if os.path.exists(lock_dir):
+            os.rmdir(lock_dir)
         # Inform control helm that the no mission is currently running.
         with open("/home/software/cuauv/workspaces/worktrees/master/control/"
                 "control_helm2/activity/mission.csv", 'w') as f:
